pyproject.toml
[tool.poetry]
name = "document-chat-bot"
version = "0.1.0"
description = ""
authors = ["justchillinghere <41962623+justchillinghere@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.5"
langchain = "^0.0.285"
pypdf = "^3.16.0"
chromadb = "^0.4.9"
openai = "^0.28.0"
tiktoken = "^0.4.0"
python-dotenv = "^1.0.0"
unstructured = "^0.10.15"
aiogram = "^3.1.1"
sqlalchemy = "^2.0.21"
psycopg2-binary = "^2.9.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
